+++ cmd/vault-secrets-webhook/main.go
@@ -1,4 +1,4 @@
-// Copyright © 2019 Banzai Cloud
+// Copyright © 2020 Banzai Cloud
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -19,26 +19,21 @@ import (
 	"fmt"
 	"net/http"
 	"strconv"
-	"strings"
-
-	internal "github.com/banzaicloud/bank-vaults/internal/configuration"
 
 	vaultapi "github.com/hashicorp/vault/api"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
-	log "github.com/sirupsen/logrus"
+	"github.com/sirupsen/logrus"
 	whhttp "github.com/slok/kubewebhook/pkg/http"
 	"github.com/slok/kubewebhook/pkg/observability/metrics"
 	whcontext "github.com/slok/kubewebhook/pkg/webhook/context"
 	"github.com/slok/kubewebhook/pkg/webhook/mutating"
 	"github.com/spf13/viper"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/errors"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	kubeVer "k8s.io/apimachinery/pkg/version"
 	"k8s.io/client-go/kubernetes"
 	kubernetesConfig "sigs.k8s.io/controller-runtime/pkg/client/config"
 
@@ -46,389 +41,75 @@ import (
 	"github.com/banzaicloud/bank-vaults/pkg/sdk/vault"
 )
 
-var vaultAgentConfig = `
-pid_file = "/tmp/pidfile"
-exit_after_auth = true
-
-auto_auth {
-	method "kubernetes" {
-		mount_path = "auth/%s"
-		config = {
-			role = "%s"
-		}
-	}
-
-	sink "file" {
-		config = {
-			path = "/vault/.vault-token"
-		}
-	}
-}`
-
-type mutatingWebhook struct {
-	k8sClient kubernetes.Interface
-	registry  registry.ImageRegistry
+// VaultConfig represents vault options
+type VaultConfig struct {
+	Addr                        string
+	Role                        string
+	Path                        string
+	SkipVerify                  string
+	TLSSecret                   string
+	UseAgent                    bool
+	VaultEnvDaemon              bool
+	TransitKeyID                string
+	TransitPath                 string
+	CtConfigMap                 string
+	CtImage                     string
+	CtOnce                      bool
+	CtImagePullPolicy           corev1.PullPolicy
+	CtShareProcess              bool
+	CtShareProcessDefault       string
+	CtCPU                       resource.Quantity
+	CtMemory                    resource.Quantity
+	PspAllowPrivilegeEscalation bool
+	IgnoreMissingSecrets        string
+	VaultEnvPassThrough         string
+	ConfigfilePath              string
+	MutateConfigMap             bool
+	EnableJSONLog               string
+	AgentConfigMap              string
+	AgentOnce                   bool
+	AgentShareProcess           bool
+	AgentShareProcessDefault    string
+	AgentCPU                    resource.Quantity
+	AgentMemory                 resource.Quantity
+	AgentImage                  string
+	AgentImagePullPolicy        corev1.PullPolicy
 }
 
-// If the original Pod contained a Volume "vault-tls", for example Vault instances provisioned by the Operator
-// we need to handle that edge case and choose another name for the vault-tls volume for accessing Vault with TLS.
-func hasTLSVolume(volumes []corev1.Volume) bool {
-	for _, volume := range volumes {
-		if volume.Name == "vault-tls" {
-			return true
-		}
-	}
-	return false
+func init() {
+	viper.SetDefault("vault_image", "vault:latest")
+	viper.SetDefault("vault_image_pull_policy", string(corev1.PullIfNotPresent))
+	viper.SetDefault("vault_env_image", "banzaicloud/vault-env:latest")
+	viper.SetDefault("vault_env_image_pull_policy", string(corev1.PullIfNotPresent))
+	viper.SetDefault("vault_ct_image", "hashicorp/consul-template:0.19.6-dev-alpine")
+	viper.SetDefault("vault_addr", "https://vault:8200")
+	viper.SetDefault("vault_skip_verify", "false")
+	viper.SetDefault("vault_path", "kubernetes")
+	viper.SetDefault("vault_role", "")
+	viper.SetDefault("vault_tls_secret", "")
+	viper.SetDefault("vault_agent", "false")
+	viper.SetDefault("vault_env_daemon", "false")
+	viper.SetDefault("vault_ct_share_process_namespace", "")
+	viper.SetDefault("psp_allow_privilege_escalation", "false")
+	viper.SetDefault("vault_ignore_missing_secrets", "false")
+	viper.SetDefault("vault_env_passthrough", "")
+	viper.SetDefault("mutate_configmap", "false")
+	viper.SetDefault("tls_cert_file", "")
+	viper.SetDefault("tls_private_key_file", "")
+	viper.SetDefault("listen_address", ":8443")
+	viper.SetDefault("telemetry_listen_address", "")
+	viper.SetDefault("default_image_pull_secret", "")
+	viper.SetDefault("default_image_pull_secret_namespace", "")
+	viper.SetDefault("default_image_pull_docker_config_json_key", corev1.DockerConfigJsonKey)
+	viper.SetDefault("registry_skip_verify", "false")
+	viper.SetDefault("debug", "false")
+	viper.SetDefault("enable_json_log", "false")
+	viper.SetDefault("vault_agent_share_process_namespace", "")
+	viper.AutomaticEnv()
 }
 
-func hasPodSecurityContextRunAsUser(p *corev1.PodSecurityContext) bool {
-	return p.RunAsUser != nil
-}
-
-func getServiceAccountMount(containers []corev1.Container) (serviceAccountMount corev1.VolumeMount) {
-mountSearch:
-	for _, container := range containers {
-		for _, mount := range container.VolumeMounts {
-			if mount.MountPath == "/var/run/secrets/kubernetes.io/serviceaccount" {
-				serviceAccountMount = mount
-				break mountSearch
-			}
-		}
-	}
-	return serviceAccountMount
-}
-
-func getInitContainers(originalContainers []corev1.Container, podSecurityContext *corev1.PodSecurityContext, vaultConfig internal.VaultConfig, initContainersMutated bool, containersMutated bool, containerEnvVars []corev1.EnvVar, containerVolMounts []corev1.VolumeMount) []corev1.Container {
-	var containers = []corev1.Container{}
-
-	if vaultConfig.UseAgent || vaultConfig.CtConfigMap != "" {
-		serviceAccountMount := getServiceAccountMount(originalContainers)
-
-		containerVolMounts = append(containerVolMounts, serviceAccountMount, corev1.VolumeMount{
-			Name:      "vault-agent-config",
-			MountPath: "/vault/agent/",
-		})
-
-		runAsUser := int64(100)
-		securityContext := &corev1.SecurityContext{
-			RunAsUser:                &runAsUser,
-			AllowPrivilegeEscalation: &vaultConfig.PspAllowPrivilegeEscalation,
-		}
-
-		containers = append(containers, corev1.Container{
-			Name:            "vault-agent",
-			Image:           vaultConfig.AgentImage,
-			ImagePullPolicy: vaultConfig.AgentImagePullPolicy,
-			SecurityContext: securityContext,
-			Command:         []string{"vault", "agent", "-config=/vault/agent/config.hcl"},
-			Env:             containerEnvVars,
-			VolumeMounts:    containerVolMounts,
-			Resources: corev1.ResourceRequirements{
-				Limits: corev1.ResourceList{
-					corev1.ResourceCPU:    resource.MustParse("50m"),
-					corev1.ResourceMemory: resource.MustParse("64Mi"),
-				},
-			},
-		})
-	}
-
-	if initContainersMutated || containersMutated {
-		containers = append(containers, corev1.Container{
-			Name:            "copy-vault-env",
-			Image:           viper.GetString("vault_env_image"),
-			ImagePullPolicy: corev1.PullPolicy(viper.GetString("vault_env_image_pull_policy")),
-			Command:         []string{"sh", "-c", "cp /usr/local/bin/vault-env /vault/"},
-			VolumeMounts: []corev1.VolumeMount{
-				{
-					Name:      "vault-env",
-					MountPath: "/vault/",
-				},
-			},
-
-			SecurityContext: getSecurityContext(podSecurityContext, vaultConfig),
-			Resources: corev1.ResourceRequirements{
-				Limits: corev1.ResourceList{
-					corev1.ResourceCPU:    resource.MustParse("50m"),
-					corev1.ResourceMemory: resource.MustParse("64Mi"),
-				},
-			},
-		})
-	}
-
-	return containers
-}
-
-func getContainers(vaultConfig internal.VaultConfig, containerEnvVars []corev1.EnvVar, containerVolMounts []corev1.VolumeMount) []corev1.Container {
-	var containers = []corev1.Container{}
-	securityContext := &corev1.SecurityContext{
-		AllowPrivilegeEscalation: &vaultConfig.PspAllowPrivilegeEscalation,
-	}
-
-	if vaultConfig.CtShareProcess {
-		securityContext = &corev1.SecurityContext{
-			AllowPrivilegeEscalation: &vaultConfig.PspAllowPrivilegeEscalation,
-			Capabilities: &corev1.Capabilities{
-				Add: []corev1.Capability{
-					"SYS_PTRACE",
-				},
-			},
-		}
-	}
-
-	containerVolMounts = append(containerVolMounts, corev1.VolumeMount{
-		Name:      "ct-secrets",
-		MountPath: "/vault/secrets",
-	}, corev1.VolumeMount{
-		Name:      "vault-env",
-		MountPath: "/home/consul-template",
-	}, corev1.VolumeMount{
-		Name:      "ct-configmap",
-		MountPath: "/vault/ct-config/config.hcl",
-		ReadOnly:  true,
-		SubPath:   "config.hcl",
-	},
-	)
-
-	var ctCommandString []string
-	if vaultConfig.CtOnce {
-		ctCommandString = []string{"-config", "/vault/ct-config/config.hcl", "-once"}
-	} else {
-		ctCommandString = []string{"-config", "/vault/ct-config/config.hcl"}
-	}
-
-	containers = append(containers, corev1.Container{
-		Name:            "consul-template",
-		Image:           vaultConfig.CtImage,
-		Args:            ctCommandString,
-		ImagePullPolicy: vaultConfig.CtImagePullPolicy,
-		SecurityContext: securityContext,
-		Env:             containerEnvVars,
-		VolumeMounts:    containerVolMounts,
-		Resources: corev1.ResourceRequirements{
-			Limits: corev1.ResourceList{
-				corev1.ResourceCPU:    vaultConfig.CtCPU,
-				corev1.ResourceMemory: vaultConfig.CtMemory,
-			},
-		},
-	})
-
-	return containers
-}
-
-func getAgentContainers(originalContainers []corev1.Container, vaultConfig internal.VaultConfig, containerEnvVars []corev1.EnvVar, containerVolMounts []corev1.VolumeMount) []corev1.Container {
-	var containers = []corev1.Container{}
-	securityContext := &corev1.SecurityContext{
-		AllowPrivilegeEscalation: &vaultConfig.PspAllowPrivilegeEscalation,
-		Capabilities: &corev1.Capabilities{
-			Add: []corev1.Capability{
-				"IPC_LOCK",
-			},
-		},
-	}
-
-	if vaultConfig.AgentShareProcess {
-		securityContext = &corev1.SecurityContext{
-			AllowPrivilegeEscalation: &vaultConfig.PspAllowPrivilegeEscalation,
-			Capabilities: &corev1.Capabilities{
-				Add: []corev1.Capability{
-					"SYS_PTRACE",
-					"IPC_LOCK",
-				},
-			},
-		}
-	}
-
-	serviceAccountMount := getServiceAccountMount(originalContainers)
-
-	containerVolMounts = append(containerVolMounts, serviceAccountMount, corev1.VolumeMount{
-		Name:      "agent-secrets",
-		MountPath: "/vault/secrets",
-	}, corev1.VolumeMount{
-		Name:      "agent-configmap",
-		MountPath: "/vault/config/config.hcl",
-		ReadOnly:  true,
-		SubPath:   "config.hcl",
-	},
-	)
-
-	var agentCommandString []string
-	if vaultConfig.AgentOnce {
-		agentCommandString = []string{"agent", "-config", "/vault/config/config.hcl", "-once"}
-	} else {
-		agentCommandString = []string{"agent", "-config", "/vault/config/config.hcl"}
-	}
-
-	containers = append(containers, corev1.Container{
-		Name:            "vault-agent",
-		Image:           vaultConfig.AgentImage,
-		Args:            agentCommandString,
-		ImagePullPolicy: vaultConfig.AgentImagePullPolicy,
-		SecurityContext: securityContext,
-		Env:             containerEnvVars,
-		VolumeMounts:    containerVolMounts,
-		Resources: corev1.ResourceRequirements{
-			Limits: corev1.ResourceList{
-				corev1.ResourceCPU:    vaultConfig.AgentCPU,
-				corev1.ResourceMemory: vaultConfig.AgentMemory,
-			},
-		},
-	})
-
-	return containers
-}
-
-func getSecurityContext(podSecurityContext *corev1.PodSecurityContext, vaultConfig internal.VaultConfig) *corev1.SecurityContext {
-	if hasPodSecurityContextRunAsUser(podSecurityContext) {
-		return &corev1.SecurityContext{
-			RunAsUser:                podSecurityContext.RunAsUser,
-			AllowPrivilegeEscalation: &vaultConfig.PspAllowPrivilegeEscalation,
-		}
-	}
-
-	return &corev1.SecurityContext{
-		AllowPrivilegeEscalation: &vaultConfig.PspAllowPrivilegeEscalation,
-	}
-}
-
-func getVolumes(existingVolumes []corev1.Volume, agentConfigMapName string, vaultConfig internal.VaultConfig, logger *log.Logger) []corev1.Volume {
-	logger.Debugf("Add generic volumes to podspec")
-
-	volumes := []corev1.Volume{
-		{
-			Name: "vault-env",
-			VolumeSource: corev1.VolumeSource{
-				EmptyDir: &corev1.EmptyDirVolumeSource{
-					Medium: corev1.StorageMediumMemory,
-				},
-			},
-		},
-	}
-
-	if vaultConfig.UseAgent || vaultConfig.CtConfigMap != "" {
-		logger.Debugf("Add vault agent volumes to podspec")
-		volumes = append(volumes, corev1.Volume{
-			Name: "vault-agent-config",
-			VolumeSource: corev1.VolumeSource{
-				ConfigMap: &corev1.ConfigMapVolumeSource{
-					LocalObjectReference: corev1.LocalObjectReference{
-						Name: agentConfigMapName,
-					},
-				},
-			},
-		})
-	}
-
-	if vaultConfig.TLSSecret != "" {
-		logger.Debugf("Add vault TLS volume to podspec")
-
-		volumeName := "vault-tls"
-		if hasTLSVolume(existingVolumes) {
-			volumeName = "vault-env-tls"
-		}
-
-		volumes = append(volumes, corev1.Volume{
-			Name: volumeName,
-			VolumeSource: corev1.VolumeSource{
-				Secret: &corev1.SecretVolumeSource{
-					SecretName: vaultConfig.TLSSecret,
-				},
-			},
-		})
-	}
-	if vaultConfig.CtConfigMap != "" {
-		logger.Debugf("Add consul template volumes to podspec")
-
-		defaultMode := int32(420)
-		volumes = append(volumes,
-			corev1.Volume{
-				Name: "ct-secrets",
-				VolumeSource: corev1.VolumeSource{
-					EmptyDir: &corev1.EmptyDirVolumeSource{
-						Medium: corev1.StorageMediumMemory,
-					},
-				},
-			},
-			corev1.Volume{
-				Name: "ct-configmap",
-				VolumeSource: corev1.VolumeSource{
-					ConfigMap: &corev1.ConfigMapVolumeSource{
-						LocalObjectReference: corev1.LocalObjectReference{
-							Name: vaultConfig.CtConfigMap,
-						},
-						DefaultMode: &defaultMode,
-						Items: []corev1.KeyToPath{
-							{
-								Key:  "config.hcl",
-								Path: "config.hcl",
-							},
-						},
-					},
-				},
-			})
-	}
-
-	if vaultConfig.AgentConfigMap != "" {
-		logger.Debugf("Add vault-agent volumes to podspec")
-
-		defaultMode := int32(420)
-		volumes = append(volumes,
-			corev1.Volume{
-				Name: "agent-secrets",
-				VolumeSource: corev1.VolumeSource{
-					EmptyDir: &corev1.EmptyDirVolumeSource{
-						Medium: corev1.StorageMediumMemory,
-					},
-				},
-			},
-			corev1.Volume{
-				Name: "agent-configmap",
-				VolumeSource: corev1.VolumeSource{
-					ConfigMap: &corev1.ConfigMapVolumeSource{
-						LocalObjectReference: corev1.LocalObjectReference{
-							Name: vaultConfig.AgentConfigMap,
-						},
-						DefaultMode: &defaultMode,
-						Items: []corev1.KeyToPath{
-							{
-								Key:  "config.hcl",
-								Path: "config.hcl",
-							},
-						},
-					},
-				},
-			})
-	}
-
-	return volumes
-}
-
-func (mw *mutatingWebhook) vaultSecretsMutator(ctx context.Context, obj metav1.Object) (bool, error) {
-	switch v := obj.(type) {
-	case *corev1.Pod:
-		return false, mw.mutatePod(v, parseVaultConfig(obj), whcontext.GetAdmissionRequest(ctx).Namespace, whcontext.IsAdmissionRequestDryRun(ctx))
-	case *corev1.Secret:
-		if _, ok := obj.GetAnnotations()["vault.security.banzaicloud.io/vault-addr"]; ok {
-			return false, mutateSecret(v, parseVaultConfig(obj))
-		}
-		return false, nil
-	case *corev1.ConfigMap:
-		if _, ok := obj.GetAnnotations()["vault.security.banzaicloud.io/mutate-configmap"]; ok {
-			return false, mutateConfigMap(v, parseVaultConfig(obj))
-		}
-		return false, nil
-
-	case *unstructured.Unstructured:
-		if _, ok := obj.GetAnnotations()["vault.security.banzaicloud.io/vault-addr"]; ok {
-			return false, mutateObject(v, parseVaultConfig(obj))
-		}
-		return false, nil
-	default:
-		return false, nil
-	}
-}
-
-func parseVaultConfig(obj metav1.Object) internal.VaultConfig {
-	var vaultConfig internal.VaultConfig
+func parseVaultConfig(obj metav1.Object) VaultConfig {
+	var vaultConfig VaultConfig
 	annotations := obj.GetAnnotations()
 
 	if val, ok := annotations["vault.security.banzaicloud.io/vault-addr"]; ok {
@@ -615,29 +296,35 @@ func parseVaultConfig(obj metav1.Object) internal.VaultConfig {
 	return vaultConfig
 }
 
-func getConfigMapForVaultAgent(pod *corev1.Pod, vaultConfig internal.VaultConfig) *corev1.ConfigMap {
-	ownerReferences := pod.GetOwnerReferences()
-	name := pod.GetName()
-	// If we have no name we are probably part of some controller,
-	// try to get the name of the owner controller.
-	if name == "" {
-		if len(ownerReferences) > 0 {
-			if strings.Contains(ownerReferences[0].Name, "-") {
-				generateNameSlice := strings.Split(ownerReferences[0].Name, "-")
-				name = strings.Join(generateNameSlice[:len(generateNameSlice)-1], "-")
-			} else {
-				name = ownerReferences[0].Name
-			}
+type mutatingWebhook struct {
+	k8sClient kubernetes.Interface
+	registry  registry.ImageRegistry
+	logger    logrus.FieldLogger
+}
+
+func (mw *mutatingWebhook) vaultSecretsMutator(ctx context.Context, obj metav1.Object) (bool, error) {
+	vaultConfig := parseVaultConfig(obj)
+	switch v := obj.(type) {
+	case *corev1.Pod:
+		return false, mw.mutatePod(v, vaultConfig, whcontext.GetAdmissionRequest(ctx).Namespace, whcontext.IsAdmissionRequestDryRun(ctx))
+	case *corev1.Secret:
+		if _, ok := obj.GetAnnotations()["vault.security.banzaicloud.io/vault-addr"]; ok {
+			return false, mw.mutateSecret(v, vaultConfig)
 		}
-	}
-	return &corev1.ConfigMap{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:            name + "-vault-agent-config",
-			OwnerReferences: ownerReferences,
-		},
-		Data: map[string]string{
-			"config.hcl": fmt.Sprintf(vaultAgentConfig, vaultConfig.Path, vaultConfig.Role),
-		},
+		return false, nil
+	case *corev1.ConfigMap:
+		if _, ok := obj.GetAnnotations()["vault.security.banzaicloud.io/mutate-configmap"]; ok {
+			return false, mw.mutateConfigMap(v, vaultConfig)
+		}
+		return false, nil
+
+	case *unstructured.Unstructured:
+		if _, ok := obj.GetAnnotations()["vault.security.banzaicloud.io/vault-addr"]; ok {
+			return false, mw.mutateObject(v, vaultConfig)
+		}
+		return false, nil
+	default:
+		return false, nil
 	}
 }
 
@@ -703,10 +390,6 @@ func (mw *mutatingWebhook) lookForEnvFrom(envFrom []corev1.EnvFromSource, ns str
 	return envVars, nil
 }
 
-func hasVaultPrefix(value string) bool {
-	return strings.HasPrefix(value, "vault:") || strings.HasPrefix(value, ">>vault:")
-}
-
 func (mw *mutatingWebhook) lookForValueFrom(env corev1.EnvVar, ns string) (*corev1.EnvVar, error) {
 	if env.ValueFrom.ConfigMapKeyRef != nil {
 		data, err := mw.getDataFromConfigmap(env.ValueFrom.ConfigMapKeyRef.Name, ns)
@@ -743,7 +426,7 @@ func (mw *mutatingWebhook) lookForValueFrom(env corev1.EnvVar, ns string) (*core
 	return nil, nil
 }
 
-func (mw *mutatingWebhook) mutateContainers(containers []corev1.Container, podSpec *corev1.PodSpec, vaultConfig internal.VaultConfig, ns string) (bool, error) {
+func (mw *mutatingWebhook) mutateContainers(containers []corev1.Container, podSpec *corev1.PodSpec, vaultConfig VaultConfig, ns string) (bool, error) {
 	mutated := false
 
 	for i, container := range containers {
@@ -896,37 +579,7 @@ func (mw *mutatingWebhook) mutateContainers(containers []corev1.Container, podSp
 	return mutated, nil
 }
 
-func addSecretsVolToContainers(vaultConfig internal.VaultConfig, containers []corev1.Container, logger *log.Logger) {
-	for i, container := range containers {
-		logger.Debugf("Add secrets VolumeMount to container %s", container.Name)
-
-		container.VolumeMounts = append(container.VolumeMounts, []corev1.VolumeMount{
-			{
-				Name:      "ct-secrets",
-				MountPath: vaultConfig.ConfigfilePath,
-			},
-		}...)
-
-		containers[i] = container
-	}
-}
-
-func addAgentSecretsVolToContainers(vaultConfig internal.VaultConfig, containers []corev1.Container, logger *log.Logger) {
-	for i, container := range containers {
-		logger.Debugf("Add secrets VolumeMount to container %s", container.Name)
-
-		container.VolumeMounts = append(container.VolumeMounts, []corev1.VolumeMount{
-			{
-				Name:      "agent-secrets",
-				MountPath: vaultConfig.ConfigfilePath,
-			},
-		}...)
-
-		containers[i] = container
-	}
-}
-
-func newVaultClient(vaultConfig internal.VaultConfig) (*vault.Client, error) {
+func newVaultClient(vaultConfig VaultConfig) (*vault.Client, error) {
 	clientConfig := vaultapi.DefaultConfig()
 	clientConfig.Address = vaultConfig.Addr
 
@@ -955,203 +608,11 @@ func newK8SClient() (kubernetes.Interface, error) {
 	return kubernetes.NewForConfig(kubeConfig)
 }
 
-func (mw *mutatingWebhook) mutatePod(pod *corev1.Pod, vaultConfig internal.VaultConfig, ns string, dryRun bool) error {
-	logger.Debugf("Successfully connected to the API")
-
-	initContainersMutated, err := mw.mutateContainers(pod.Spec.InitContainers, &pod.Spec, vaultConfig, ns)
-	if err != nil {
-		return err
-	}
-
-	if initContainersMutated {
-		logger.Debugf("Successfully mutated pod init containers")
-	} else {
-		logger.Debugf("No pod init containers were mutated")
-	}
-
-	containersMutated, err := mw.mutateContainers(pod.Spec.Containers, &pod.Spec, vaultConfig, ns)
-	if err != nil {
-		return err
-	}
-
-	if containersMutated {
-		logger.Debugf("Successfully mutated pod containers")
-	} else {
-		logger.Debugf("No pod containers were mutated")
-	}
-
-	containerEnvVars := []corev1.EnvVar{
-		{
-			Name:  "VAULT_ADDR",
-			Value: vaultConfig.Addr,
-		},
-		{
-			Name:  "VAULT_SKIP_VERIFY",
-			Value: vaultConfig.SkipVerify,
-		},
-	}
-	containerVolMounts := []corev1.VolumeMount{
-		{
-			Name:      "vault-env",
-			MountPath: "/vault/",
-		},
-	}
-	if vaultConfig.TLSSecret != "" {
-		mountPath := "/vault/tls/ca.crt"
-		volumeName := "vault-tls"
-		if hasTLSVolume(pod.Spec.Volumes) {
-			mountPath = "/vault-env/tls/ca.crt"
-			volumeName = "vault-env-tls"
-		}
-
-		containerEnvVars = append(containerEnvVars, corev1.EnvVar{
-			Name:  "VAULT_CACERT",
-			Value: mountPath,
-		})
-		containerVolMounts = append(containerVolMounts, corev1.VolumeMount{
-			Name:      volumeName,
-			MountPath: mountPath,
-			SubPath:   "ca.crt",
-		})
-	}
-
-	if initContainersMutated || containersMutated || vaultConfig.CtConfigMap != "" || vaultConfig.AgentConfigMap != "" {
-		var agentConfigMapName string
-
-		if vaultConfig.UseAgent || vaultConfig.CtConfigMap != "" {
-			configMap := getConfigMapForVaultAgent(pod, vaultConfig)
-			agentConfigMapName = configMap.Name
-
-			if !dryRun {
-				_, err := mw.k8sClient.CoreV1().ConfigMaps(ns).Create(configMap)
-				if err != nil {
-					if errors.IsAlreadyExists(err) {
-						_, err = mw.k8sClient.CoreV1().ConfigMaps(ns).Update(configMap)
-						if err != nil {
-							return err
-						}
-					} else {
-						return err
-					}
-				}
-			}
-		}
-
-		pod.Spec.InitContainers = append(getInitContainers(pod.Spec.Containers, pod.Spec.SecurityContext, vaultConfig, initContainersMutated, containersMutated, containerEnvVars, containerVolMounts), pod.Spec.InitContainers...)
-		logger.Debugf("Successfully appended pod init containers to spec")
-
-		pod.Spec.Volumes = append(pod.Spec.Volumes, getVolumes(pod.Spec.Volumes, agentConfigMapName, vaultConfig, logger)...)
-		logger.Debugf("Successfully appended pod spec volumes")
-	}
-
-	if vaultConfig.CtConfigMap != "" {
-		logger.Debugf("Consul Template config found")
-
-		addSecretsVolToContainers(vaultConfig, pod.Spec.Containers, logger)
-
-		if vaultConfig.CtShareProcessDefault == "empty" {
-			logger.Debugf("Test our Kubernetes API Version and make the final decision on enabling ShareProcessNamespace")
-			apiVersion, _ := mw.k8sClient.Discovery().ServerVersion()
-			versionCompared := kubeVer.CompareKubeAwareVersionStrings("v1.12.0", apiVersion.String())
-			logger.Debugf("Kubernetes API version detected: %s", apiVersion.String())
-
-			if versionCompared >= 0 {
-				vaultConfig.CtShareProcess = true
-			} else {
-				vaultConfig.CtShareProcess = false
-			}
-		}
-
-		if vaultConfig.CtShareProcess {
-			logger.Debugf("Detected shared process namespace")
-			shareProcessNamespace := true
-			pod.Spec.ShareProcessNamespace = &shareProcessNamespace
-		}
-		pod.Spec.Containers = append(getContainers(vaultConfig, containerEnvVars, containerVolMounts), pod.Spec.Containers...)
-
-		logger.Debugf("Successfully appended pod containers to spec")
-	}
-
-	if vaultConfig.AgentConfigMap != "" {
-		logger.Debugf("Vault Agent config found")
-
-		addAgentSecretsVolToContainers(vaultConfig, pod.Spec.Containers, logger)
-
-		if vaultConfig.AgentShareProcessDefault == "empty" {
-			logger.Debugf("Test our Kubernetes API Version and make the final decision on enabling ShareProcessNamespace")
-			apiVersion, _ := mw.k8sClient.Discovery().ServerVersion()
-			versionCompared := kubeVer.CompareKubeAwareVersionStrings("v1.12.0", apiVersion.String())
-			logger.Debugf("Kubernetes API version detected: %s", apiVersion.String())
-
-			if versionCompared >= 0 {
-				vaultConfig.AgentShareProcess = true
-			} else {
-				vaultConfig.AgentShareProcess = false
-			}
-		}
-
-		if vaultConfig.AgentShareProcess {
-			logger.Debugf("Detected shared process namespace")
-			shareProcessNamespace := true
-			pod.Spec.ShareProcessNamespace = &shareProcessNamespace
-		}
-		pod.Spec.Containers = append(getAgentContainers(pod.Spec.Containers, vaultConfig, containerEnvVars, containerVolMounts), pod.Spec.Containers...)
-
-		logger.Debugf("Successfully appended pod containers to spec")
-	}
-	return nil
-}
-
-func init() {
-	viper.SetDefault("vault_image", "vault:latest")
-	viper.SetDefault("vault_image_pull_policy", string(corev1.PullIfNotPresent))
-	viper.SetDefault("vault_env_image", "banzaicloud/vault-env:latest")
-	viper.SetDefault("vault_env_image_pull_policy", string(corev1.PullIfNotPresent))
-	viper.SetDefault("vault_ct_image", "hashicorp/consul-template:0.19.6-dev-alpine")
-	viper.SetDefault("vault_addr", "https://vault:8200")
-	viper.SetDefault("vault_skip_verify", "false")
-	viper.SetDefault("vault_path", "kubernetes")
-	viper.SetDefault("vault_role", "")
-	viper.SetDefault("vault_tls_secret", "")
-	viper.SetDefault("vault_agent", "false")
-	viper.SetDefault("vault_env_daemon", "false")
-	viper.SetDefault("vault_ct_share_process_namespace", "")
-	viper.SetDefault("psp_allow_privilege_escalation", "false")
-	viper.SetDefault("vault_ignore_missing_secrets", "false")
-	viper.SetDefault("vault_env_passthrough", "")
-	viper.SetDefault("mutate_configmap", "false")
-	viper.SetDefault("tls_cert_file", "")
-	viper.SetDefault("tls_private_key_file", "")
-	viper.SetDefault("listen_address", ":8443")
-	viper.SetDefault("telemetry_listen_address", "")
-	viper.SetDefault("default_image_pull_secret", "")
-	viper.SetDefault("default_image_pull_secret_namespace", "")
-	viper.SetDefault("default_image_pull_docker_config_json_key", corev1.DockerConfigJsonKey)
-	viper.SetDefault("registry_skip_verify", "false")
-	viper.SetDefault("debug", "false")
-	viper.SetDefault("enable_json_log", "false")
-	viper.SetDefault("vault_agent_share_process_namespace", "")
-	viper.AutomaticEnv()
-
-	logger = log.New()
-	// Add additional fields to all log messages
-	logger.AddHook(&GlobalHook{})
-
-	if viper.GetBool("enable_json_log") {
-		logger.SetFormatter(&log.JSONFormatter{})
-	}
-
-	if viper.GetBool("debug") {
-		logger.SetLevel(log.DebugLevel)
-		logger.Debug("Debug mode enabled")
-	}
-}
-
 func healthzHandler(w http.ResponseWriter, r *http.Request) {
 	w.WriteHeader(200)
 }
 
-func handlerFor(config mutating.WebhookConfig, mutator mutating.MutatorFunc, recorder metrics.Recorder, logger *log.Logger) http.Handler {
+func handlerFor(config mutating.WebhookConfig, mutator mutating.MutatorFunc, recorder metrics.Recorder, logger logrus.FieldLogger) http.Handler {
 	webhook, err := mutating.NewWebhook(config, mutator, nil, recorder, logger)
 	if err != nil {
 		logger.Fatalf("error creating webhook: %s", err)
@@ -1160,35 +621,34 @@ func handlerFor(config mutating.WebhookConfig, mutator mutating.MutatorFunc, rec
 	return whhttp.MustHandlerFor(webhook)
 }
 
-var logger *log.Logger
-
-// GlobalHook struct used for adding additional fields to the log
-type GlobalHook struct {
-}
-
-// Levels returning all log levels
-func (h *GlobalHook) Levels() []log.Level {
-	return log.AllLevels
-}
-
-// Fire adding the additional fields to all log entries
-func (h *GlobalHook) Fire(e *log.Entry) error {
-	e.Data["app"] = "vault-secrets-webhook"
-	return nil
-}
-
-func serveMetrics(addr string) {
-	logger.Infof("Telemetry on http://%s", addr)
+func (mw *mutatingWebhook) serveMetrics(addr string) {
+	mw.logger.Infof("Telemetry on http://%s", addr)
 
 	mux := http.NewServeMux()
 	mux.Handle("/metrics", promhttp.Handler())
 	err := http.ListenAndServe(addr, mux)
 	if err != nil {
-		logger.Fatalf("error serving telemetry: %s", err)
+		mw.logger.Fatalf("error serving telemetry: %s", err)
 	}
 }
 
 func main() {
+	var logger logrus.FieldLogger
+	{
+		log := logrus.New()
+
+		if viper.GetBool("enable_json_log") {
+			log.SetFormatter(&logrus.JSONFormatter{})
+		}
+
+		if viper.GetBool("debug") {
+			log.SetLevel(logrus.DebugLevel)
+			log.Debug("Debug mode enabled")
+		}
+
+		logger = log.WithField("app", "vault-secrets-webhook")
+	}
+
 	k8sClient, err := newK8SClient()
 	if err != nil {
 		logger.Fatalf("error creating k8s client: %s", err)
@@ -1197,6 +657,7 @@ func main() {
 	mutatingWebhook := mutatingWebhook{
 		k8sClient: k8sClient,
 		registry:  registry.NewRegistry(),
+		logger:    logger,
 	}
 
 	mutator := mutating.MutatorFunc(mutatingWebhook.vaultSecretsMutator)
@@ -1222,7 +683,7 @@ func main() {
 
 	if len(telemetryAddress) > 0 {
 		// Serving metrics without TLS on separated address
-		go serveMetrics(telemetryAddress)
+		go mutatingWebhook.serveMetrics(telemetryAddress)
 	} else {
 		mux.Handle("/metrics", promhttp.Handler())
 	}
