+++ internal/configuration/vaultConfig.go
@@ -13,43 +13,3 @@
 // limitations under the License.
 
 package configuration
-
-import (
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
-)
-
-// VaultConfig represents vault options
-type VaultConfig struct {
-	Addr                        string
-	Role                        string
-	Path                        string
-	SkipVerify                  string
-	TLSSecret                   string
-	UseAgent                    bool
-	VaultEnvDaemon              bool
-	TransitKeyID                string
-	TransitPath                 string
-	CtConfigMap                 string
-	CtImage                     string
-	CtOnce                      bool
-	CtImagePullPolicy           corev1.PullPolicy
-	CtShareProcess              bool
-	CtShareProcessDefault       string
-	CtCPU                       resource.Quantity
-	CtMemory                    resource.Quantity
-	PspAllowPrivilegeEscalation bool
-	IgnoreMissingSecrets        string
-	VaultEnvPassThrough         string
-	ConfigfilePath              string
-	MutateConfigMap             bool
-	EnableJSONLog               string
-	AgentConfigMap              string
-	AgentOnce                   bool
-	AgentShareProcess           bool
-	AgentShareProcessDefault    string
-	AgentCPU                    resource.Quantity
-	AgentMemory                 resource.Quantity
-	AgentImage                  string
-	AgentImagePullPolicy        corev1.PullPolicy
-}
